+++ packages/tron-wallet-snap/src/ui/confirmation/views/ConfirmTransactionRequest/ConfirmTransactionRequest.tsx
@@ -47,7 +47,8 @@ export const ConfirmTransactionRequest = ({
 
   /**
    * Only disable confirm button upon first load (FetchStatus.Loading)
-   * as opposed to subsequent loads (FetchStatus.Fetching)
+   * as opposed to subsequent loads (FetchStatus.Fetching), and whenever
+   * the simulation has failed.
    */
   const shouldDisableConfirmButton =
     scanFetchStatus === FetchStatus.Loading ||
@@ -88,7 +89,7 @@ export const ConfirmTransactionRequest = ({
     <Container>
       <Box>
         {/* Security Alert */}
-        {preferences.useSecurityAlerts ? (
+        {preferences.useSecurityAlerts || scan?.error ? (
           <TransactionAlert
             scanFetchStatus={scanFetchStatus}
             validation={scan?.validation ?? null}